+++ src/lib/pages/interact-contract/components/InteractionTypeSwitch.tsx
@@ -76,7 +76,7 @@ export const InteractionTypeSwitch = <
           h="22px"
           position="absolute"
           borderRadius="2px"
-          backgroundColor="primary.dark"
+          backgroundColor="primary.darker"
           animate={{
             left: activeIndex === 0 ? "4px" : "100px",
           }}